向量表达从double[]切换为float[]

+++ src/main/java/io/github/ompc/erniebot4j/embedding/Embedding.java
@@ -1,6 +1,6 @@
 package io.github.ompc.erniebot4j.embedding;
 
-public record Embedding(int index, double[] vectors) implements Comparable<Embedding> {
+public record Embedding(int index, float[] vectors) implements Comparable<Embedding> {
 
     @Override
     public int compareTo(Embedding o) {
